refactor: 特徴量抽出メソッドのファイル名を変更
