+++ backend/src/routers/estimate_sleep.py
@@ -4,7 +4,9 @@ from pandera.typing import DataFrame
 
 from src.schemas.estimate_sleep import EstimateSleepRequest
 from src.schemas.extract_steps import StepCountDFSchema
-from src.usecases.estimate_sleep.feature_of_late_night import create_feature_value
+from src.usecases.estimate_sleep.feature_of_late_night_usecase import (
+    create_feature_value,
+)
 from src.usecases.estimate_sleep.step_clustering_usecase import step_clustering
 
 router = APIRouter(prefix="/estimate-sleep", tags=["estimate-sleep"])
